backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enterprise-mcp-agent-backend"
version = "0.1.0"
description = "Enterprise MCP Agent System - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "openai>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis[hiredis]>=5.0.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "nemoguardrails>=0.11.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "langsmith>=0.1.0",
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["UP042"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E402", "E501", "F841", "N806", "N815", "E741"]
"app/agent/prompts.py" = ["E501"]
"app/agent/tools.py" = ["F841"]
"app/agent/nodes/**/*.py" = ["N814", "N817", "E501"]
"app/api/v1/mcp_servers.py" = ["N815"]
"app/guardrails/rails.py" = ["N806"]
"app/middleware/error_handler.py" = ["E741"]
"app/services/chat_service.py" = ["F841"]
"app/graphs/**/*.py" = ["N806"]
"app/**/schemas.py" = ["N815"]
"mcp_servers/**/*.py" = ["N814", "N817", "F841"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
